requirements.txt
flask
flask-pymongo
flask-jwt-extended
python-dotenv
werkzeug
